# adds performance sample