--- a/examples/performance-sample/build.gradle
+++ b/examples/performance-sample/build.gradle
@@ -1,0 +1,11 @@
+plugins {
+    // this plugin lets you create properties files
+    // for multiple environments... like dev, qa, prod
+    id 'net.saliman.properties' version '1.4.6'
+
+    // this is the data hub framework gradle plugin
+    // it includes ml-gradle. This plugin is what lets you
+    // run DHF (Data Hub Framework) tasks from the
+    // command line
+    id 'com.marklogic.ml-data-hub' version '2.0.0'
+}