closes #492

All the details are in the README -- this should fulfill all the requirements in #492.

I used the geonames cities5000, this results in 45669 records, loaded first as XML and agains as JSON. (For some reason, "cities5000" contains 45677 records, and there's a handful of bad-character errors.) There's many similar geonames collections ranging dramatically in size, so it'd be easy to switch to a different source for shorter or longer testing.

On my MBP after running `hubInit` in an empty directory, `doAll` takes ~6 minutes (including a full, first-time `mlDeploy`). Here's the timing output with `-Pprofile`:

```
BUILD SUCCESSFUL in 6m 8s
21 actionable tasks: 21 executed
Task timings:
     2.742s  :hubPreInstallCheck
     0.001s  :mlDeleteModuleTimestampsFile
     0.004s  :mlPrepareRestApiDependencies
    88.791s  :mlDeployApp
     0.000s  :mlPostDeploy
     0.000s  :mlDeploy
     0.004s  :createJsonEntity
     0.002s  :createInputJsonFlow
     0.001s  :createXmlEntity
     0.002s  :createInputXmlFlow
     0.000s  :createEntityInput
     0.005s  :createHarmonizeJsonSjs
     0.003s  :createHarmonizeJsonXqy
     0.002s  :createHarmonizeXmlSjs
     0.002s  :createHarmonizeXmlXqy
     0.000s  :createHarmonizeFlows
     0.418s  :getInputData
     3.595s  :mlLoadModules
    52.553s  :loadJson
    52.685s  :loadXml
     0.003s  :loadInputData
    51.452s  :runHarmonizeJsonSjs
    29.802s  :runHarmonizeJsonXqy
    55.021s  :runHarmonizeXmlSjs
    30.660s  :runHarmonizeXmlXqy
     0.000s  :allHarmonizeFlows
     0.000s  :doAll
```

If you'd like to track timing over time (perhaps in a CI job), the [build-time-tracker-plugin](https://github.com/passy/build-time-tracker-plugin) seems superior to the add-hoc timing class in this PR, as it can append timing reports to a log.

And if the need arises to profile the data-hub implementation itself, there are many additional options that should be considered. A couple that stand out are the [gradle profiler](https://github.com/gradle/gradle-profiler), which has hooks for many different profiling packages, and [gradle build scans](https://guides.gradle.org/performance/#build-scan-performance), which appears to require (free) use of Gradle Cloud Services.